requirements.txt
fastapi==0.101.1
pydantic==1.10.6
pymongo==4.3.3
python-dotenv==1.0.0
uvicorn==0.21.0
